backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
sqlalchemy
requests
pydantic
ollama
python-magic
aiofiles
PyPDF2
Pillow